オフライン時とオンライン時の処理ををifで分岐させた。

+++ lib/history/history_model.dart
@@ -1,4 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:posture_correction/utils.dart';
 
@@ -32,6 +33,10 @@ class HistoryModel extends ChangeNotifier {
 
   //メモをアップデート
   Future updateTitle(Data data) async {
+    final connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      return;
+    }
     await FirebaseFirestore.instance
         .collection("users")
         .doc(Utils.userId)
@@ -42,15 +47,23 @@ class HistoryModel extends ChangeNotifier {
             : {
                 "memo": "",
               });
+
+    await fetchData();
   }
 
   //計測データ消去
   Future delete(Data data) async {
-    return await FirebaseFirestore.instance
+    final connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      throw ("通信状態をご確認ください");
+    }
+    await FirebaseFirestore.instance
         .collection("users")
         .doc(Utils.userId)
         .collection("measurements")
         .doc(data.documentID)
         .delete();
+
+    await fetchData();
   }
 }
